+++ src/store/actionCreator.js
@@ -1,0 +1,13 @@
+import {actionType} from './actionType';
+
+export const ActionCreator = {
+  changeCity: (city) => ({
+    type: actionType.changeCity,
+    payload: city
+  }),
+
+  loadOffers: (offers) => ({
+    type: actionType.loadOffers,
+    payload: offers
+  }),
+};
