+++ src/store/actionType.js
@@ -1,0 +1,4 @@
+export const actionType = {
+  loadOffers: `LOAD_OFFERS`,
+  changeCity: `CHANGE_CITY`,
+};
