+++ src/store/initialState.js
@@ -1,0 +1,4 @@
+export const initialState = {
+  city: ``,
+  offers: [],
+};
